pengembalian done

+++ pages/staff/peminjaman/view/index.php
@@ -35,12 +35,17 @@ require_once __DIR__ . "/../action/show.php";
                         <td><?= ucwords($item['status']) ?></td>
                         <td>
                             <div class="flex items-center gap-2 justify-center">
+                                <a href="<?= BASE_URL ?>/pages/staff/peminjaman/action/kembalikan.php?id=<?= $item['id'] ?>" class="p-1 w-8 bg-blue-500 text-white rounded-lg">
+                                    <i class="fa-solid fa-check"></i>
+                                </a>
                                 <a href="<?= BASE_URL ?>/pages/staff/peminjaman/view/edit.php?id=<?= $item['id'] ?>" class="p-1 w-8 bg-yellow-500 text-white rounded-lg">
                                     <i class="fa-solid fa-pen-to-square"></i>
                                 </a>
-                                <a href="<?= BASE_URL ?>/pages/staff/peminjaman/action/delete.php?id=<?= $item['id'] ?>" class="p-1 w-8 bg-red-500 text-white rounded-lg">
-                                    <i class="fa-solid fa-trash-can"></i>
-                                </a>
+                                <?php if ($item['status'] != 'dikembalikan'): ?>
+                                    <a href="<?= BASE_URL ?>/pages/staff/peminjaman/action/delete.php?id=<?= $item['id'] ?>" class="p-1 w-8 bg-red-500 text-white rounded-lg">
+                                        <i class="fa-solid fa-trash-can"></i>
+                                    </a>
+                                <?php endif; ?>
                             </div>
                         </td>
                     </tr>
